RST-6454 refactor testing project tree for better readability and structure
