+++ tests/end_to_end_tests/journey_1/uploads/section.py
@@ -1,7 +1,6 @@
 from playwright.async_api import Page
-from tests.helpers.e2e_assert_helpers import AssertHelpers
-from tests.helpers.e2e_page_helpers import PageHelpers
-import tests.end_to_end_tests.journey_1.data as data
+from tests.end_to_end_tests.helpers.e2e_assert_helpers import AssertHelpers
+from tests.end_to_end_tests.helpers.e2e_page_helpers import PageHelpers
 import tests.end_to_end_tests.journey_1.uploads.page_name_change_documents as page_name_change_documents
 import tests.end_to_end_tests.journey_1.uploads.page_marriage_documents as page_marriage_documents
 import tests.end_to_end_tests.journey_1.uploads.page_overseas_documents as page_overseas_documents
